Recording code moved to seperate file
